move git functions to same file
